split constraint builders code
